--- youtube.py
@@ -1,17 +1,0 @@
-from langchain.document_loaders.generic import GenericLoader
-from langchain.document_loaders.parsers import OpenAIWhisperParser
-from langchain.document_loaders.blob_loaders.youtube_audio import YoutubeAudioLoader
-
-
-# ! pip install yt_dlp
-# ! pip install pydub
-
-
-url="https://www.youtube.com/watch?v=jGwO_UgTS7I"
-save_dir="docs/youtube/"
-loader = GenericLoader(
-    YoutubeAudioLoader([url],save_dir),
-    OpenAIWhisperParser()
-)
-docs = loader.load()
-print(docs[0].page_content[0:500])
